print event
