+++ ops/aws/canary/lambda/cmd/alarm_slack_handler/main.go
@@ -1,61 +1,24 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/secretsmanager"
 	"github.com/filecoin-project/bacalhau/ops/aws/canary/pkg/logger"
-	"os"
 )
 
-// Type to hold slack webhooks retrieved from secret manager
-type slackWebhooksType struct {
-	AlarmOk        string `json:"alarmOk"`
-	AlarmTriggered string `json:"alarmTriggered"`
-}
-
 // variable to store the slack webhooks to retrieve them once and reuse them across recent invocations
-var slackWebhooks = mustGetWebhookSecret()
-
-func mustGetWebhookSecret() slackWebhooksType {
-	secretName := os.Getenv("SLACK_WEBHOOK_SECRET_NAME")
-
-	//Create a Secrets Manager client
-	sess, err := session.NewSession()
-	if err != nil {
-		panic(err)
-	}
-
-	svc := secretsmanager.New(sess)
-	input := &secretsmanager.GetSecretValueInput{
-		SecretId: aws.String(secretName),
-	}
-
-	result, err := svc.GetSecretValue(input)
-	if err != nil {
-		panic(err)
-	}
-
-	newSlackWebhooks := slackWebhooksType{}
-	err = json.Unmarshal([]byte(*result.SecretString), &newSlackWebhooks)
-	if err != nil {
-		panic(err)
-	}
-
-	return newSlackWebhooks
-}
+var slackWebhooks slackWebhooksType
 
 func init() {
 	logger.SetupCWLogger()
+	slackWebhooks = mustGetWebhookSecret()
 }
 
-func handle(event events.CloudWatchAlarmSNSPayload) error {
-	fmt.Println(event)
-	fmt.Printf("Webhooks: %+v\n", slackWebhooks)
+func handle(event events.SNSEvent) error {
+	for _, record := range event.Records {
+		fmt.Printf("SNS record: %+v\n", record)
+	}
 	return nil
 }
